# docs/openapi/schema.yaml
openapi: 3.0.3
info:
  title: BudgetWiseBackend API
  version: 1.0.0
  description: OpenAPI-документация backend-части прогрессивного веб-приложения для
    управления личными финансами.
paths:
...
  /api/v1/users/me/:
    get:
      operationId: users_me_retrieve
      summary: Получить данные текущего пользователя
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
          description: ''
    patch:
      operationId: users_me_partial_update
      summary: Частично обновить данные текущего пользователя
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCurrentUserRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCurrentUserRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCurrentUserRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
          description: ''
...
components:
  schemas:
...
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
          title: Имя пользователя
          description: Обязательное поле. Не более 150 символов. Только буквы, цифры
            и символы @/./+/-/_.
        email:
          type: string
          format: email
          readOnly: true
        first_name:
          type: string
          title: Имя
          maxLength: 150
        last_name:
          type: string
          title: Фамилия
          maxLength: 150
        role:
          type: string
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
          title: Активный
          description: Отметьте, если пользователь должен считаться активным. Уберите
            эту отметку вместо удаления учётной записи.
        date_joined:
          type: string
          format: date-time
          readOnly: true
          title: Дата регистрации
      required:
      - date_joined
      - email
      - id
      - is_active
      - role
      - username
...
    PatchedCurrentUserRequest:
      type: object
      properties:
        first_name:
          type: string
          title: Имя
          maxLength: 150
        last_name:
          type: string
          title: Фамилия
          maxLength: 150
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...